requirements.txt
discord.py
python-dotenv
black
# discord-py-slash-command